+++ libraries/effect/src/main/java/androidx/media3/effect/SpeedChangeShaderProgram.java
@@ -15,24 +15,95 @@
  */
 package androidx.media3.effect;
 
+import androidx.media3.common.C;
 import androidx.media3.common.GlObjectsProvider;
 import androidx.media3.common.GlTextureInfo;
+import androidx.media3.common.audio.SpeedProvider;
 import androidx.media3.common.util.UnstableApi;
 
-/** Applies a speed change by updating the frame timestamps. */
+/**
+ * Applies the speed changes specified in a {@link SpeedProvider} change by updating the frame
+ * timestamps.
+ *
+ * <p>Does not support seeking in effects previewing.
+ */
 @UnstableApi
 /* package */ final class SpeedChangeShaderProgram extends PassthroughShaderProgram {
 
-  private final float speed;
+  private final OffsetSpeedProvider speedProvider;
 
-  public SpeedChangeShaderProgram(float speed) {
+  private long lastSpeedChangeInputTimeUs;
+  private long lastSpeedChangeOutputTimeUs;
+
+  public SpeedChangeShaderProgram(SpeedProvider speedProvider) {
     super();
-    this.speed = speed;
+    this.speedProvider = new OffsetSpeedProvider(speedProvider);
+    lastSpeedChangeInputTimeUs = C.TIME_UNSET;
+    lastSpeedChangeOutputTimeUs = C.TIME_UNSET;
   }
 
   @Override
   public void queueInputFrame(
       GlObjectsProvider glObjectsProvider, GlTextureInfo inputTexture, long presentationTimeUs) {
-    super.queueInputFrame(glObjectsProvider, inputTexture, (long) (presentationTimeUs / speed));
+    long outputPresentationTimeUs;
+    if (lastSpeedChangeInputTimeUs == C.TIME_UNSET) {
+      outputPresentationTimeUs = presentationTimeUs;
+      lastSpeedChangeInputTimeUs = presentationTimeUs;
+      lastSpeedChangeOutputTimeUs = outputPresentationTimeUs;
+      speedProvider.setOffset(presentationTimeUs);
+    } else {
+      long nextSpeedChangeInputTimeUs =
+          speedProvider.getNextSpeedChangeTimeUs(lastSpeedChangeInputTimeUs);
+      while (nextSpeedChangeInputTimeUs != C.TIME_UNSET
+          && nextSpeedChangeInputTimeUs <= presentationTimeUs) {
+        lastSpeedChangeOutputTimeUs =
+            getOutputTimeUs(
+                nextSpeedChangeInputTimeUs, speedProvider.getSpeed(lastSpeedChangeInputTimeUs));
+        lastSpeedChangeInputTimeUs = nextSpeedChangeInputTimeUs;
+        nextSpeedChangeInputTimeUs =
+            speedProvider.getNextSpeedChangeTimeUs(lastSpeedChangeInputTimeUs);
+      }
+      outputPresentationTimeUs =
+          getOutputTimeUs(presentationTimeUs, speedProvider.getSpeed(presentationTimeUs));
+    }
+    super.queueInputFrame(glObjectsProvider, inputTexture, outputPresentationTimeUs);
+  }
+
+  @Override
+  public void signalEndOfCurrentInputStream() {
+    super.signalEndOfCurrentInputStream();
+    lastSpeedChangeInputTimeUs = C.TIME_UNSET;
+    lastSpeedChangeOutputTimeUs = C.TIME_UNSET;
+  }
+
+  private long getOutputTimeUs(long inputTimeUs, float speed) {
+    return (long)
+        (lastSpeedChangeOutputTimeUs + (inputTimeUs - lastSpeedChangeInputTimeUs) / speed);
+  }
+
+  private static class OffsetSpeedProvider implements SpeedProvider {
+
+    private final SpeedProvider speedProvider;
+
+    private long offset;
+
+    public OffsetSpeedProvider(SpeedProvider speedProvider) {
+      this.speedProvider = speedProvider;
+    }
+
+    public void setOffset(long offset) {
+      this.offset = offset;
+    }
+
+    @Override
+    public float getSpeed(long timeUs) {
+      return speedProvider.getSpeed(timeUs - offset);
+    }
+
+    @Override
+    public long getNextSpeedChangeTimeUs(long timeUs) {
+      long nextSpeedChangeTimeUs = speedProvider.getNextSpeedChangeTimeUs(timeUs - offset);
+      return nextSpeedChangeTimeUs == C.TIME_UNSET ? C.TIME_UNSET : offset + nextSpeedChangeTimeUs;
+    }
   }
 }